+++ src/ewokscore/graph/graph_io.py
@@ -1,7 +1,5 @@
 import logging
-from difflib import get_close_matches
 from typing import Dict
-from typing import Iterable
 from typing import Iterator
 from typing import List
 from typing import Mapping
@@ -14,6 +12,7 @@ from .. import missing_data
 from ..node import NodeIdType
 from ..node import get_node_label
 from ..task import Task
+from ..utils import build_close_match_report
 from .analysis import end_nodes
 from .analysis import start_nodes
 
@@ -50,23 +49,6 @@ def update_default_inputs(
                 default_inputs.append(input_item)
         else:
             node_attrs["default_inputs"] = [input_item]
-
-
-def suggest_close_match(word: str, possibilities: Iterable[str]) -> Optional[str]:
-    """Return the closest match from possibilities, or None if none found."""
-    close_match = get_close_matches(word, list(possibilities), n=1, cutoff=0.5)
-    return close_match[0] if close_match else None
-
-
-def build_close_match_report(words: Iterable[str], possibilities: Iterable[str]) -> str:
-    """Return an error report with close match suggestions for each word."""
-    possibilities = list(possibilities)
-    error_report = ""
-    for word in words:
-        suggestion = suggest_close_match(word, possibilities)
-        if suggestion:
-            error_report += f"\n  @ You provided '{suggestion}'. Did you mean '{word}'?"
-    return error_report
 
 
 def parse_inputs(
